requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.54.3
langchain==0.3.7
langchain-openai==0.2.8
langchain-core==0.3.15
langgraph==0.2.45
Werkzeug==3.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
